+++ src/test/java/tobyspring/splearn/SplearnTestConfiguration.java
@@ -3,7 +3,7 @@ package tobyspring.splearn;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import tobyspring.splearn.application.member.required.EmailSender;
-import tobyspring.splearn.domain.MemberFixture;
+import tobyspring.splearn.domain.member.MemberFixture;
 import tobyspring.splearn.domain.member.PasswordEncoder;
 
 @TestConfiguration
