+++ src/test/java/tobyspring/splearn/application/member/provided/MemberFinderTest.java
@@ -10,7 +10,7 @@ import org.springframework.context.annotation.Import;
 import org.springframework.transaction.annotation.Transactional;
 import tobyspring.splearn.SplearnTestConfiguration;
 import tobyspring.splearn.domain.member.Member;
-import tobyspring.splearn.domain.MemberFixture;
+import tobyspring.splearn.domain.member.MemberFixture;
 
 @SpringBootTest
 @Transactional
