+++ src/test/java/tobyspring/splearn/application/member/provided/MemberRegisterManualTest.java
@@ -19,7 +19,7 @@ import tobyspring.splearn.application.member.required.EmailSender;
 import tobyspring.splearn.application.member.required.MemberRepository;
 import tobyspring.splearn.domain.shared.Email;
 import tobyspring.splearn.domain.member.Member;
-import tobyspring.splearn.domain.MemberFixture;
+import tobyspring.splearn.domain.member.MemberFixture;
 import tobyspring.splearn.domain.member.MemberStatus;
 
 class MemberRegisterManualTest {
