+++ src/test/java/tobyspring/splearn/application/member/required/MemberRepositoryTest.java
@@ -2,8 +2,8 @@ package tobyspring.splearn.application.member.required;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static tobyspring.splearn.domain.MemberFixture.createMemberRegisterRequest;
-import static tobyspring.splearn.domain.MemberFixture.createPasswordEncoder;
+import static tobyspring.splearn.domain.member.MemberFixture.createMemberRegisterRequest;
+import static tobyspring.splearn.domain.member.MemberFixture.createPasswordEncoder;
 
 import jakarta.persistence.EntityManager;
 import org.junit.jupiter.api.Test;
