+++ src/test/java/tobyspring/splearn/domain/member/MemberTest.java
@@ -1,16 +1,13 @@
-package tobyspring.splearn.domain;
+package tobyspring.splearn.domain.member;
 
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static tobyspring.splearn.domain.MemberFixture.createMemberRegisterRequest;
-import static tobyspring.splearn.domain.MemberFixture.createPasswordEncoder;
+import static tobyspring.splearn.domain.member.MemberFixture.createMemberRegisterRequest;
+import static tobyspring.splearn.domain.member.MemberFixture.createPasswordEncoder;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import tobyspring.splearn.domain.member.Member;
-import tobyspring.splearn.domain.member.MemberStatus;
-import tobyspring.splearn.domain.member.PasswordEncoder;
 
 class MemberTest {
     Member member;
