+++ src/test/java/tobyspring/splearn/domain/shared/EmailTest.java
@@ -1,11 +1,10 @@
-package tobyspring.splearn.domain;
+package tobyspring.splearn.domain.shared;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.Test;
-import tobyspring.splearn.domain.shared.Email;
 
 class EmailTest {
